feat: add Cursors/Room switch in settings and sorting by different parameters

+++ app/src/main/java/com/example/rsschooltask4/viewmodel/TaskViewModel.kt
@@ -1,24 +1,42 @@
 package com.example.rsschooltask4.viewmodel
 
+import android.content.Context
+import android.content.SharedPreferences
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import androidx.preference.PreferenceManager
+import com.example.rsschooltask4.data.CursorRepository
 import com.example.rsschooltask4.data.RoomRepository
 import com.example.rsschooltask4.data.model.ItemData
+import com.example.rsschooltask4.view.MainActivity
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
-class TaskViewModel : ViewModel() {
+class TaskViewModel(context: Context) : ViewModel() {
+
+    private val roomIsOn by lazy {
+    val prefs: SharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
+        prefs.getBoolean("cursor-room_switch", false)
+    }
+
+    private val sortParameter by lazy {
+        val prefs: SharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
+        prefs.getString("list_preference", "id")
+    }
+
     private val roomRepository = RoomRepository.get()
+    private val cursorRepository = CursorRepository.get()
+
 //    lateinit var list: MutableLiveData<MutableList<ItemData>>
 
     fun addRecord(itemData: ItemData) {
         viewModelScope.launch(Dispatchers.IO) {
-            roomRepository.addItem(itemData)
+            if (roomIsOn) roomRepository.addItem(itemData) else cursorRepository.addItem(itemData)
         }
     }
 
-    val list = roomRepository.readTable()
+    val list = if (roomIsOn) roomRepository.readTable(sortParameter) else cursorRepository.readTable(sortParameter)
 
 //    fun readRecords(): MutableLiveData<MutableList<ItemData>> {
 //        viewModelScope.launch(Dispatchers.IO) {
@@ -34,13 +52,17 @@ class TaskViewModel : ViewModel() {
         thirdParameter: String
     ) {
         viewModelScope.launch(Dispatchers.IO) {
-            roomRepository.updateItem(id, firstParameter, secondParameter, thirdParameter)
+            if (roomIsOn) {
+                roomRepository.updateItem(id, firstParameter, secondParameter, thirdParameter)
+            } else {
+                cursorRepository.updateItem(id, firstParameter, secondParameter, thirdParameter)
+            }
         }
     }
 
     fun deleteRecord(id: Long) {
         viewModelScope.launch(Dispatchers.IO) {
-            roomRepository.deleteItem(id)
+            if (roomIsOn) roomRepository.deleteItem(id) else cursorRepository.deleteItem(id)
         }
     }
 }
